flight-sql: add GetTableTypes metadata RPC; fix build + clippy drift
This session's additions (platform Phase 02 follow-up):
- krishiv-flight-sql: implement the previously-unimplemented GetTableTypes
  metadata RPC (get_flight_info_table_types / do_get_table_types), emitting
  the single `TABLE` type, mirroring do_get_catalogs. Some BI tools call it
  during introspection; it formerly returned arrow-flight's default
  "unimplemented". +1 regression test (table_types_lists_table); crate
  suite 64 green. G1 was already fixed/verified 2026-07-02; this fills a
  remaining metadata hole. No Java toolchain in this env to re-run the live
  JDBC spike, so verified by unit test only.
- krishiv-api/engines.rs: fix a type mismatch in the STREAM-5 background
  checkpoint path (source_in_flight expects Vec<(String, Vec<Vec<u8>>)>;
  wrap snapshot_in_flight() with the source name, matching the stop-path
  pattern at the B-5 site) so the workspace compiles.
- krishiv-delta/trace.rs, krishiv-state/rocksdb_backend.rs: fix two clippy
  1.92 findings (needless Some+?, redundant usize cast) hit while linting.

Also folds in an in-progress parallel workstream present in the tree
(API-1 Flight SQL result-size cap, STREAM-5 checkpoint, related api/
engine-core/scheduler changes), committed together per maintainer direction.

Known separate issue (pre-existing, already on origin, NOT addressed here):
the full-workspace build fails in krishiv-connectors lakehouse/dml.rs:484-486
(CONN-4 version-hint block) under feature unification — write_version_hint
is feature-gated out and metadata_location() is a stale iceberg-0.9.1 API
reference. Needs an iceberg-API fix; out of scope for this change.

+++ crates/krishiv-state/src/rocksdb_backend.rs
@@ -50,7 +50,7 @@ impl RocksDbStateBackend {
             .ok()
             .and_then(|v| v.parse::<usize>().ok())
             .unwrap_or(64);
-        opts.set_write_buffer_size(write_buffer_mb as usize * 1024 * 1024);
+        opts.set_write_buffer_size(write_buffer_mb * 1024 * 1024);
         opts.set_max_write_buffer_number(3);
         let max_open_files = std::env::var("KRISHIV_ROCKSDB_MAX_OPEN_FILES")
             .ok()
